refactor(cockpit): move health tests to owner module
